pyproject.toml
[project]
name = "radioglobe"
version = "0.1.0"
description = "RadioGlobe is an internet radio player originally by Jude Pullen (Hey Jude), Donald Robson and Peter Milne."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.13",
    "lgpio>=0.2.2.0",
    "python-vlc>=3.0.21203",
    "requests>=2.32.4",
    "rpi-lgpio>=0.6",
    "smbus>=1.1.post2",
    "spidev>=3.7",
]

[dependency-groups]
dev = [
    "pytest>=8.4.0",
]

[tool.ruff]
line-length = 100
indent-width = 4
